Move coordinate transformations to geometry.utils
